moved errors from toc/entry to toc/exceptions
